+++ src/fireblocks-sdk.ts
@@ -139,7 +139,9 @@ import {
     ScreeningSupportedAssetResponse,
     ScreeningSupportedProviders,
     RegisterAssetResponse,
-    UnspentInputsResponse, AssetPriceResponse,
+    UnspentInputsResponse,
+    ContractAddressResponseDto,
+    AssetPriceResponse,
 } from "./types";
 import { AxiosProxyConfig, AxiosResponse, InternalAxiosRequestConfig } from "axios";
 import { PIIEncryption } from "./pii-client";
@@ -1958,6 +1960,16 @@ export class FireblocksSDK {
     }
 
     /**
+     * Get contract's address by blockchain base assetId and the transaction hash
+     * @param baseAssetId
+     * @param txHash
+     *
+     * @returns {ContractAddressResponseDto}
+     */
+    public async getContractAddress(baseAssetId: string, txHash: string): Promise<ContractAddressResponseDto> {
+        return await this.apiClient.issueGetRequest(`/v1/contract_interactions/base_asset_id/${baseAssetId}/tx_hash/${txHash}`);
+    }
+    /**
      * Get contract by blockchain base assetId and contract address
      * @param baseAssetId
      * @param templateId
@@ -1976,7 +1988,7 @@ export class FireblocksSDK {
      * @returns {ContractAbiResponseDto}
      */
     public async getContractAbi(baseAssetId: string, contractAddress: string): Promise<ContractAbiResponseDto> {
-        return await this.apiClient.issueGetRequest(`/v1/contract_interactions/base_asset_id/${baseAssetId}/contract_address/${contractAddress}/abi`);
+        return await this.apiClient.issueGetRequest(`/v1/contract_interactions/base_asset_id/${baseAssetId}/contract_address/${contractAddress}/functions`);
     }
 
     /**