+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
+#### [v5.27.2](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.27.1...v5.27.2)
+
+> 20 June 2024
+
+- feat: set ncw wallet required algorithms [`#281`](https://github.com/fireblocks/fireblocks-sdk-js/pull/281)
+
+#### [v5.27.1](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.27.0...v5.27.1)
+
+> 20 June 2024
+
+- functions instead of abi [`#280`](https://github.com/fireblocks/fireblocks-sdk-js/pull/280)
+
+#### [v5.27.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.26.0...v5.27.0)
+
+> 17 June 2024
+
+- Add algorithm id to asset response [`#271`](https://github.com/fireblocks/fireblocks-sdk-js/pull/271)
+- Fetch address by chain and tx hash [`#279`](https://github.com/fireblocks/fireblocks-sdk-js/pull/279)
+
 #### [v5.26.0](https://github.com/fireblocks/fireblocks-sdk-js/compare/v5.25.1...v5.26.0)
 
 > 28 May 2024